Squashed commit for version 0.0.2 (merging feature-v0.0.2 branch)  - Updating ChangeLog with version 0.0.2  - Adding VERSION file.  - Small documentation update:     1. Added warning that we may not support latest versions of frameworks     2. Added some words about python dependencies.  - Changing logging progress from logging.info to regular print every 10 benchmarks.  - Fixing bug that misses some errors in Caffe2 log files.  - Adding script that demonstrates advanced experimenter usage such as:       1. Multi-GPU benchmarks       2. Enabling/disabling experiments  - Adding more comments for a tutorial script.  - Fixing bug that counts disabled experiments twice.  - Adding a user-friendly message that tells how to gracefully terminate benchmarking process.  - Getting rid of 'bc' tool to do floating point math in bash since it's not installed by default in many OSs. Switched to python.  - Adding functionality to fracefully shutdown benchmarking process. One needs to send a SIGUSR1 signal:      kill -USR1 PID, where PID is a process ID that is printed by a experimenter (launcher pid: PID)  - Adding support to disable certain functionality if a particular python module is not available.  - Adding python dependencies. They are all for now optional - if something is missing, corresponding functionality will be disabled.  - Catching no pandas module could be imported and disabling certain functionality on collecting system info in this case.  - Use exp.framework as a launcher key name of exp.framework_family is missing. Fixing bugs related to new parameter naming convention.  - Fixing bugs related to new parameter naming convention. Making certain parameters optional (added checks for parameter existence).  - Updating reports modules to version 0.0.2 to use new parameter naming convention.  - Fixing typo in parameter description.  - Updating documentation - adding new parameter description sections and adding references to sysinfo, resource monitor and data modules.  - Updating/enhancing/fixing bugs for tutorial scripts.  - Script that automatically generates parameter documentation.  - Fixing TensorRT tutorial examples.  - Adding checks for TensorRT package when building docker image.  - This is a first commit in for version 0.0.2. Major updates that are not compatible with previous version:       1. A new naming schema for input/output parameters - more structured and unified.    Also:       1. Several modifications have been made to support half precision in Caffe2 and MXNet framework.       2. Documentation added for input data, system info collection, data precision specification and reporting.       3. Documentation is partially updated to reflect changes in input/output parameters.       4. Refactored parameter help module. Help messages are now defined in same files where parameters are.       5. Parameters now have types.

+++ docker/build.sh
@@ -128,6 +128,20 @@ for dockerfile_dir in "$@"; do
     # If we are building tensorrt docker, we need to copy source of benchmark
     # project to docker context folder
     if [ "$name" == "tensorrt" ]; then
+        # One special thing about building TensorRT images is that we need to have
+        # a TensorRT package in a docke file folder. The name of that file must be
+        # specified in a 'versions' file - so we need verify user has copied this
+        # file there.
+        if [ "${version}XXX" == "XXX" ]; then
+            logwarn "Will not build TensorRT ($dockerfile_dir) because name of a package is missing in 'versions' file."
+            continue
+        fi
+        if [ ! -f "$dockerfile_dir/$version" ]; then
+            logwarn "Will not build TensorRT ($dockerfile_dir) because TensorRT package ($dockerfile_dir/$version) not found."
+            logwarn "You must copy corresponding package (most likely, *.deb file) into that folder."
+            logwarn "You can get it from NVIDIA developer site."
+            continue
+        fi
         rm -rf -- $dockerfile_dir/tensorrt       # Delete if old version exists
         cp -r ../src/tensorrt  $dockerfile_dir   # Copy project
     fi
